converted doubles to decimals

+++ POSTerminalGui/CashPaymentMethod.cs
@@ -12,15 +12,15 @@ namespace POSTerminalGui
             return "Cash $" + amountPaid;
         }
 
-        public override void TakePayment(double amountDue, out double amountStillDue)
+        public override void TakePayment(decimal amountDue, out decimal amountStillDue)
         {
            //For cash, ask the amount tendered and provide change.
            //For check, get the check number.
            //For credit, get the credit card number, expiration, and CW.
 
            Console.WriteLine("Please entered the amount tendered: ");
-           double amountTendered = 0;
-           while (!double.TryParse(Console.ReadLine(), out amountTendered) ||
+           decimal amountTendered = 0;
+           while (!decimal.TryParse(Console.ReadLine(), out amountTendered) ||
                 amountTendered <= 0)
             {
                 Console.WriteLine("I'm sorry I didn't understand.  Please try again.");
@@ -28,7 +28,7 @@ namespace POSTerminalGui
 
            if (amountTendered >= amountDue)
             {
-                double change = amountTendered - amountDue;
+                decimal change = amountTendered - amountDue;
                 //amountPaid = amountDue;
                 amountPaid = amountTendered;
                 //Console.WriteLine("Here is your change: $" + change);
